add spacing between image item, image thumb zoom in animation

+++ core_lib/src/main/java/com/ninhhk/faster/data/store/LruDiskStore.java
@@ -264,6 +264,10 @@ public class LruDiskStore extends DiskStore {
 
     @Override
     public void clear() {
+        if (! config.isUseDiskCache()){
+            return;
+        }
+
         try {
             diskLruCache.delete();
         } catch (IOException e) {
